cleanup metadata functions
