+++ src/serlo-editor-integration/types/editor-plugins.ts
@@ -204,6 +204,7 @@ export interface EditorTemplateExerciseGroupDocument {
     uuid?: number
     trashed?: boolean
     unrevisedRevisions?: number
+    license?: License
   }
 }
